# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS REST API",
        "description": "Human Resource Management System API Documentation",
        "contact": {
            "email": "admin@hrms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/expense-categories": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get all expense categories",
                "operationId": "d5f0cf9eac09e862a20236b4e93a9830",
                "responses": {
                    "200": {
                        "description": "List of expense categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "expense_name": {
                                                        "type": "string",
                                                        "example": "Office Supplies"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create new expense category",
                "operationId": "2356ab1a8a84632ba1531d05a5ea5951",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "expense_name"
                                ],
                                "properties": {
                                    "expense_name": {
                                        "type": "string",
                                        "example": "Office Supplies"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/expense-categories/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get expense category details",
                "operationId": "def3bfe7e50c7868ca58fb24b06ebc70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Update expense category",
                "operationId": "389cc6d01bb62a4a32d5cc23c288e0ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "expense_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Delete expense category",
                "operationId": "b31c0b8732933354b570430e65fd86ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get all expense entries",
                "operationId": "b1b77dc505b7fd85a1ae7f1b3880f8af",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of expenses"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create new expense entry",
                "operationId": "7eca81a4560f60faba9c02b3071397ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "expense_category",
                                    "amount",
                                    "expense_date",
                                    "payment_type"
                                ],
                                "properties": {
                                    "expense_category": {
                                        "type": "string",
                                        "example": "Office Supplies"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-15"
                                    },
                                    "payment_type": {
                                        "description": "1=Cash, 2=Bank",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bank_name": {
                                        "type": "string",
                                        "example": "Bank Name (required if payment_type=2)"
                                    },
                                    "remark": {
                                        "type": "string",
                                        "example": "Purchase of office supplies"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/expenses/statement": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get expense statement (by category)",
                "operationId": "ec078a8e420d5fb0366d44894b6e47f3",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense statement"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}